Small fixes in Scribo.
	* demo/src/mainwindow.cc: Do not always resize pics.

	* filter/object_groups_small.hh: Add missing traces.

	* preprocessing/split_bg_fg.hh: Fix type ambiguities.

	* src/table/extract.cc,
	* src/text/grouping/group_from_double_link_filter_non_aligned.cc,
	* src/text/grouping/group_from_rag.cc: Cosmetic changes.

git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@4548 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ scribo/src/text/grouping/group_from_double_link_filter_non_aligned.cc
@@ -141,7 +141,7 @@ int main(int argc, char *argv[])
   // Filter non-correctly aligned grouped objects.
 
   object_links<L> filtered_links
-    = filter::object_links_non_h_aligned(text, double_link, atoi(argv[4]));
+    = filter::object_links_non_h_aligned(text, double_link, atof(argv[4]));
 
 
   text_t aligned_grouped_text
